Independant controller repositioning
- Fixed different systems controllers being positioned incorrectly after customising the position of one controller layout.
- Removed KissXML as it's no longer used.

+++ Provenance/Categories/UIImage+Scaling.m
@@ -12,7 +12,7 @@
 
 - (UIImage *)scaledImageWithMaxResolution:(NSInteger)maxResolution
 {
-	int kMaxResolution = maxResolution;
+	NSInteger kMaxResolution = maxResolution;
     
     CGImageRef imgRef = self.CGImage;
